+++ src/full/Agda/TypeChecking/Pretty/Warning.hs
@@ -165,7 +165,7 @@ prettyWarning = \case
             , fsep $ pwords "Suggestion: add a `variable Any : Set _` and replace unsolved metas by Any"
             ]
 
-    AbsurdPatternRequiresNoRHS ps -> fwords $
+    AbsurdPatternRequiresAbsentRHS ps -> fwords $
       "The right-hand side must be omitted if there " ++
       "is an absurd pattern, () or {}, in the left-hand side."